Fix polarized quirp oscillator rates
The polarized chances were reversed for vertical and horizontal quirps. Using the default highchance of 40 as an example, getPolarizedChances(true) would now return {40, 40, 10, 10}, which matches up with the quirps in QuirpOscillator.java (UP, DOWN, LEFT, RIGHT) for 80% vertical, and 20% horizontal.

I am assuming vertical means up and down, and horizontal left and right; if not this is useless, and I will look very stupid.